requirements.txt
numpy==1.17.4
tqdm==4.40.2
xarray==0.14.1
matplotlib==3.1.1
h5py==2.9.0
Cartopy==0.17.0
ipdb==0.12.3
pandas==0.25.3
library==0.0.0
madrigalWeb==3.1.12
netCDF4==1.5.3
python_dateutil==2.8.1
